refactor: User 도메인 아키텍처 개선
- repository 패키지 추가

- infrastructure 패키지에 존재하던, Generator 클래스들 support 클래스로 이동

- 도메인 패키지에서  관리하던 Generator 인터페이스 제거(추구 개선 예정)
