--- src/main/java/gravit/code/user/domain/HandleGenerator.java
@@ -1,5 +1,0 @@
-package gravit.code.user.domain;
-
-public interface HandleGenerator {
-    String generateUniqueHandle();
-}
