+++ src/main/java/gravit/code/user/infrastructure/RedisMailAuthCodeConstants.java
@@ -1,4 +1,4 @@
-package gravit.code.user.infrastructure.mail;
+package gravit.code.user.infrastructure;
 
 import lombok.experimental.UtilityClass;
 import org.springframework.data.redis.core.script.DefaultRedisScript;
