+++ src/main/java/gravit/code/user/repository/custom/UserDeletionRepository.java
@@ -1,0 +1,5 @@
+package gravit.code.user.repository.custom;
+
+public interface UserDeletionRepository {
+    void cleanUserDeletion(long userId);
+}
