+++ src/main/java/gravit/code/user/service/UserDeletionService.java
@@ -5,9 +5,11 @@ import gravit.code.global.exception.domain.CustomErrorCode;
 import gravit.code.global.exception.domain.RestApiException;
 import gravit.code.user.config.UserDeleteMailProps;
 import gravit.code.user.domain.User;
-import gravit.code.user.domain.UserRepository;
-import gravit.code.user.infrastructure.MailAuthCodeGenerator;
+import gravit.code.user.repository.UserRepository;
+import gravit.code.user.support.MailAuthCodeGenerator;
 import gravit.code.user.infrastructure.RedisUserCleanManager;
+import gravit.code.user.service.port.MailAuthCodeStore;
+import gravit.code.user.service.port.MailSender;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
