+++ src/main/java/gravit/code/user/service/port/MailAuthCodeStore.java
@@ -1,4 +1,4 @@
-package gravit.code.user.service;
+package gravit.code.user.service.port;
 
 public interface MailAuthCodeStore {
     void save(
