+++ src/test/java/gravit/code/auth/support/RandomHandleGeneratorTest.java
@@ -2,7 +2,7 @@
 //
 //import gravit.code.global.exception.domain.RestApiException;
 //import gravit.code.user.domain.UserRepository;
-//import gravit.code.user.infrastructure.RandomHandleGenerator;
+//import gravit.code.user.support.RandomHandleGenerator;
 //import org.junit.jupiter.api.Test;
 //import org.junit.jupiter.api.extension.ExtendWith;
 //import org.mockito.InjectMocks;
